day02/airflow_requirements.txt
apache-airflow==2.8.0
pandas
pyyaml
